Add bots manager
